Cache all country data
The goal here is to remove both runtime dependencies and slow
code that has be loaded. It's for performance and to just ship
less code which is always a good thing assuming things work the
same way.

I added two new cache generation commands one for lists and the
other for formatted output. This makes things much simpler at
runtime though of course all of the complexity is in pre-computing
and making sure that the pre-computed data is up-to-date.