Update cache and tests after country changes
The countries gem got updated so a few things had to be changed as well.
The key difference here is that a few translations are different and
the Czech language is no longer associated with Slovakia.

I also updated the rakefile to automatically run the outdated readme
and cache checks when run with no arguments.