Prep for v1.0.0
The only breaking change here is dropping support for Ruby 2.7. I did
that a few weeks ago and now it looks like the countries gem has done
the same thing. It was beyond end of life so that is not surprising.

Beyond that there were a bunch of small dependency updates but nothing
really worth mentioning. You can look at the diff if you're curious.